Probe the fps of the source video and specify for final video
This fixes an issue with videos with that declare
a variable frame rate but actually aren't (a
surprising amount!?), mencoder would use the
maximum rate, meaning the video in the final
video would end before the music had finished.

This will not work for videos that have truely
variable frame rates. As far as I know, mencoder
doesn't support these.

Useful environment variable that drops you into
a shell if the program end unexpectedly.