[pm3] More robust error handling
Before: `Encoder.encode()` returns str on success and str on failure, there is no indication of whether the encode is good or bad. We can guess that the encode is bad if the output file is missing or empty, but if the output is truncated we have no way of knowing.

After: Have `Encoder.encode()` return `None` on success and `Exception` (with relevant debug info) on failure. Also, check ffmpeg return code and throw an exception if it is non-zero.