[pl3] splash scale manual override
There was an issue where the splash screen 3D elements were too large and covering up the 2D elements. It turned out that the Three API changed and we needed to call a new `updateProjectionMatrix()` to get the auto-scaling to work properly, so manual scaling wasn't needed, but keeping this code around for reference just in case.